chqk8sentitygraphexporter (#189)
Add an exporter that reads content sourced from the k8sobjects receiver, makes summaries of the contents, and sends this to Cardinal's knowledge graph for use with Chip AI.

Currently only pods, secrets, and configmaps are summarized.  Secrets and configmaps have their values hashed with SHA256 with a prefix of the cluster name, kind, secret/configmap name, and the specific data item name as a prefix.  This ensures that no two secrets are trivially comparable, but change detection for a specific field can be used for Chip to make better replies.